Move `HTTPCredentials` and `TokenPayload` from `models` to `web_server.auth`
